# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Valet Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SiteSettings {
  id: number;
  siteName: string;
  heroTitle: string;
  heroSubtitle: string;
  aboutTitle: string;
  aboutText: string;
  servicesTitle: string;
  bookingTitle: string;
  bookingSubtitle: string;
  footerText: string;
  metaDescription: string;
  updatedAt: string;
}

export interface UpdateSiteSettingsBody {
  siteName?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  aboutTitle?: string;
  aboutText?: string;
  servicesTitle?: string;
  bookingTitle?: string;
  bookingSubtitle?: string;
  footerText?: string;
  metaDescription?: string;
}

export interface Section {
  id: number;
  name: string;
  label: string;
  isVisible: boolean;
  sortOrder: number;
  updatedAt: string;
}

export interface CreateSectionBody {
  name: string;
  label: string;
  isVisible?: boolean;
  sortOrder?: number;
}

export interface UpdateSectionBody {
  name?: string;
  label?: string;
  isVisible?: boolean;
  sortOrder?: number;
}

export interface ReorderSectionsBody {
  orderedIds: number[];
}

export interface ContactInfo {
  id: number;
  whatsappNumber: string;
  whatsappMessage: string;
  email: string;
  phone: string;
  address: string;
  bookingUrl: string;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  updatedAt: string;
}

export interface UpdateContactInfoBody {
  whatsappNumber?: string;
  whatsappMessage?: string;
  email?: string;
  phone?: string;
  address?: string;
  bookingUrl?: string;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
}

export interface ColorTheme {
  id: number;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  backgroundColor: string;
  textColor: string;
  headerBgColor: string;
  footerBgColor: string;
  buttonColor: string;
  buttonTextColor: string;
  linkColor: string;
  updatedAt: string;
}

export interface UpdateColorThemeBody {
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  backgroundColor?: string;
  textColor?: string;
  headerBgColor?: string;
  footerBgColor?: string;
  buttonColor?: string;
  buttonTextColor?: string;
  linkColor?: string;
}