# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Valet Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all site settings
 */
export const GetSettingsResponse = zod.object({
  id: zod.number(),
  siteName: zod.string(),
  heroTitle: zod.string(),
  heroSubtitle: zod.string(),
  aboutTitle: zod.string(),
  aboutText: zod.string(),
  servicesTitle: zod.string(),
  bookingTitle: zod.string(),
  bookingSubtitle: zod.string(),
  footerText: zod.string(),
  metaDescription: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update site settings
 */
export const UpdateSettingsBody = zod.object({
  siteName: zod.string().optional(),
  heroTitle: zod.string().optional(),
  heroSubtitle: zod.string().optional(),
  aboutTitle: zod.string().optional(),
  aboutText: zod.string().optional(),
  servicesTitle: zod.string().optional(),
  bookingTitle: zod.string().optional(),
  bookingSubtitle: zod.string().optional(),
  footerText: zod.string().optional(),
  metaDescription: zod.string().optional(),
});

export const UpdateSettingsResponse = zod.object({
  id: zod.number(),
  siteName: zod.string(),
  heroTitle: zod.string(),
  heroSubtitle: zod.string(),
  aboutTitle: zod.string(),
  aboutText: zod.string(),
  servicesTitle: zod.string(),
  bookingTitle: zod.string(),
  bookingSubtitle: zod.string(),
  footerText: zod.string(),
  metaDescription: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get all page sections
 */
export const GetSectionsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  label: zod.string(),
  isVisible: zod.boolean(),
  sortOrder: zod.number(),
  updatedAt: zod.string(),
});
export const GetSectionsResponse = zod.array(GetSectionsResponseItem);

/**
 * @summary Create a new section
 */
export const CreateSectionBody = zod.object({
  name: zod.string(),
  label: zod.string(),
  isVisible: zod.boolean().optional(),
  sortOrder: zod.number().optional(),
});

/**
 * @summary Reorder sections
 */
export const ReorderSectionsBody = zod.object({
  orderedIds: zod.array(zod.number()),
});

export const ReorderSectionsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  label: zod.string(),
  isVisible: zod.boolean(),
  sortOrder: zod.number(),
  updatedAt: zod.string(),
});
export const ReorderSectionsResponse = zod.array(ReorderSectionsResponseItem);

/**
 * @summary Get a section by ID
 */
export const GetSectionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSectionResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  label: zod.string(),
  isVisible: zod.boolean(),
  sortOrder: zod.number(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a section
 */
export const UpdateSectionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSectionBody = zod.object({
  name: zod.string().optional(),
  label: zod.string().optional(),
  isVisible: zod.boolean().optional(),
  sortOrder: zod.number().optional(),
});

export const UpdateSectionResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  label: zod.string(),
  isVisible: zod.boolean(),
  sortOrder: zod.number(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a section
 */
export const DeleteSectionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get contact information
 */
export const GetContactResponse = zod.object({
  id: zod.number(),
  whatsappNumber: zod.string(),
  whatsappMessage: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  address: zod.string(),
  bookingUrl: zod.string(),
  facebookUrl: zod.string().nullish(),
  instagramUrl: zod.string().nullish(),
  twitterUrl: zod.string().nullish(),
  updatedAt: zod.string(),
});

/**
 * @summary Update contact information
 */
export const UpdateContactBody = zod.object({
  whatsappNumber: zod.string().optional(),
  whatsappMessage: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  address: zod.string().optional(),
  bookingUrl: zod.string().optional(),
  facebookUrl: zod.string().nullish(),
  instagramUrl: zod.string().nullish(),
  twitterUrl: zod.string().nullish(),
});

export const UpdateContactResponse = zod.object({
  id: zod.number(),
  whatsappNumber: zod.string(),
  whatsappMessage: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  address: zod.string(),
  bookingUrl: zod.string(),
  facebookUrl: zod.string().nullish(),
  instagramUrl: zod.string().nullish(),
  twitterUrl: zod.string().nullish(),
  updatedAt: zod.string(),
});

/**
 * @summary Get color theme
 */
export const GetColorsResponse = zod.object({
  id: zod.number(),
  primaryColor: zod.string(),
  secondaryColor: zod.string(),
  accentColor: zod.string(),
  backgroundColor: zod.string(),
  textColor: zod.string(),
  headerBgColor: zod.string(),
  footerBgColor: zod.string(),
  buttonColor: zod.string(),
  buttonTextColor: zod.string(),
  linkColor: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update color theme
 */
export const UpdateColorsBody = zod.object({
  primaryColor: zod.string().optional(),
  secondaryColor: zod.string().optional(),
  accentColor: zod.string().optional(),
  backgroundColor: zod.string().optional(),
  textColor: zod.string().optional(),
  headerBgColor: zod.string().optional(),
  footerBgColor: zod.string().optional(),
  buttonColor: zod.string().optional(),
  buttonTextColor: zod.string().optional(),
  linkColor: zod.string().optional(),
});

export const UpdateColorsResponse = zod.object({
  id: zod.number(),
  primaryColor: zod.string(),
  secondaryColor: zod.string(),
  accentColor: zod.string(),
  backgroundColor: zod.string(),
  textColor: zod.string(),
  headerBgColor: zod.string(),
  footerBgColor: zod.string(),
  buttonColor: zod.string(),
  buttonTextColor: zod.string(),
  linkColor: zod.string(),
  updatedAt: zod.string(),
});