# lib/api-zod/src/generated/types/contactInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Valet Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInfo {
  id: number;
  whatsappNumber: string;
  whatsappMessage: string;
  email: string;
  phone: string;
  address: string;
  bookingUrl: string;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  updatedAt: string;
}