# lib/api-zod/src/generated/types/reorderSectionsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Valet Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface ReorderSectionsBody {
  orderedIds: number[];
}