# lib/api-zod/src/generated/types/updateColorThemeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Valet Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateColorThemeBody {
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  backgroundColor?: string;
  textColor?: string;
  headerBgColor?: string;
  footerBgColor?: string;
  buttonColor?: string;
  buttonTextColor?: string;
  linkColor?: string;
}