# lib/api-zod/src/generated/types/updateSiteSettingsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Valet Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSiteSettingsBody {
  siteName?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  aboutTitle?: string;
  aboutText?: string;
  servicesTitle?: string;
  bookingTitle?: string;
  bookingSubtitle?: string;
  footerText?: string;
  metaDescription?: string;
}